改了好多地方啊。。。。基本上可以跑起来了，改完了，但还没测试！！！！，前端 没改，主要是改后端。先COMMIT一下吧，浪费了不好！！还有，远程调用的部分的配置瞎写类的，在另外一个项目里面。COMMIT再说吧。

+++ Crowdfunding/src/main/java/com/crowdfunding/sjtu/Vo/RequestSerialVO.java
@@ -5,7 +5,7 @@ public class RequestSerialVO {
 	private int userid;
 	private int projectid;
 	private int shares;
-	private double price;
+	private float price;
 	/**
 	 * @return the id
 	 */
@@ -57,13 +57,13 @@ public class RequestSerialVO {
 	/**
 	 * @return the price
 	 */
-	public double getPrice() {
+	public float getPrice() {
 		return price;
 	}
 	/**
 	 * @param price the price to set
 	 */
-	public void setPrice(double price) {
+	public void setPrice(float price) {
 		this.price = price;
 	}
